fixed menu animation
