+++ src/Header/styles.module.scss
@@ -6,7 +6,7 @@
   background-color: var(--color-primary);
   position: fixed;
   width: 100%;
-  z-index: 2;
+  z-index: 1;
 
   .container {
     display: flex;
